Move brent.h to where it is used
